Merging forward calandar function into main (#4)
* Connected inst-cal-view to store & calandarDataSlice

imported useDispatch and useSelector with CalandarDataSlice into instructor-calandar-view

* Attempting to isolate the starting day of each week

* Days of Week

Instructor Calandar now shows correct days of week, created weekdays and daysofweek function to determine which day is (getDay()) map over weekDaysArr and return which day of the month it is on the current day. then the rest of the week is determined by subtracting the other indexes

* connected days of week number to weekNavSLice state

* week advance

date now advances past the month and adjusts for the correct end of month value. however, the 'next month' feature does not advance to n=2

* end of day

* better logic

updated dayOfWeek to be dayOfMonth and added better logic to return the correct day. Still working how to update the advanceMonth state

* advanceMonthAdvance

AdvanceMonthAdvance in weekNaveSlice now updates the month state properly to display current month based on advanceMonth state

* State save for Advance Month

utilized useEffect to update state of advanceMonthAdvance, which in turns updates the month

* start of month exception

some months began with a 0 first day. so changed use effect to trigger on day 2, and set end of month = 0 to end of month = 1 (which governs weeks running over the end of the month) this way, the fix doens't trigger a new month with the use effect. checked for a year and it was ok. the calandar breaks down a year or so in

* calandar advances through next month

* calandar works for one year

breaks around one year, a day ahead

* Calandar advance works

* working on last week function

* undo to next week working

* working on last week

be careful with useEffect, need to figure out when to trigger reverseMonthAdvance. right now the nextWeek system works

* Trying to simplify and debug next month. Seems to be the advance month feature is problematic. Am removing it and adding to month state the same way I do to baseDay

* Days ahead and Days behind

seperated the calandar logic into two secions, days ahead (of baseday) and days behind. now I can adjust the month differently for each senario when the month changes

* Calandar Advance Complete

Calandar works for every day of the week, adjusts for leap year, going forward. Now will working on reversing

* Reverse Month

Trying to get reverse month to work by setting all calculations to the opposite of advance month

* removed console logs and unnecessary notes

* removed unnecessary actions, reducers in weekNavSlice and instructor-calandar-view